started working on rogue & ranger

+++ src/main/java/hero/Ranger.java
@@ -36,16 +36,16 @@ public class Ranger extends Hero{
     }
 
     @Override
-    public int calcDamage() {
+    public double calcDamage() {
         int weaponDmg = 1;
         //check if weapon is equipped
         if(super.getEquipment().get(Slot.WEAPON) != null){
 
-            weaponDmg = ((Weapon)super.getEquipment().get(Slot.WEAPON)).getWeaponDamage();
+            weaponDmg = ((Weapon) super.getEquipment().get(Slot.WEAPON)).getWeaponDamage();
         }
 
-        float heroDmg = weaponDmg * ((float)  (1 + (this.heroAttributes.getDexterity()  /100)));
+        double heroDmg = weaponDmg * ((double)  (1 + (this.heroAttributes.getDexterity()  /100)));
 
-        return (int) heroDmg;
+        return heroDmg;
     }
 }
