+++ src/main/java/hero/Warrior.java
@@ -32,15 +32,15 @@ public class Warrior extends Hero{
     }
 
     @Override
-    public int calcDamage() {
+    public double calcDamage() {
         int weaponDmg = 1;
 
         //check if weapon is equipped
         if(super.getEquipment().get(Slot.WEAPON) != null) {
             weaponDmg = ((Weapon) super.getEquipment().get(Slot.WEAPON)).getWeaponDamage();
         }
-        float heroDmg = weaponDmg * ((float)  (1 + (this.heroAttributes.getStrength()  /100)));
+        double heroDmg = weaponDmg * ((double)  (1 + (this.heroAttributes.getStrength()  /100)));
 
-        return (int) heroDmg;
+        return  heroDmg;
     }
 }
